Move CeFi signin form to components/cefi
